Extract and improve test suite for `second_derivative`
* Extract tests for `second_derivative` from `Tests/test_Calculus.py` to a new dedicated test file `Tests/test_second_derivatives.py`
* Clean up duplication of tests in the old test block
* Add new tests for missing edge cases (e.g. power equal to one, large powers, negative coefficient and positive power)
* Eliminate obsolete `second_derivative` import from `Tests/test_Calculus.py`
